Implement prior sampling, error display, calibration and calibration plots

+++ harmoni_pm/common/array.py
@@ -30,7 +30,7 @@
 
 import numpy as np
 
-ARRAY_TYPE = 'float32'
+ARRAY_TYPE = 'float64'
 
 class FloatArray(np.ndarray):
     def __new__(
